+++ src/dune_copasi_md.cc
@@ -6,11 +6,17 @@
 #include <dune/copasi/common/stepper.hh>
 #include <dune/copasi/grid/mark_stripes.hh>
 #include <dune/copasi/grid/multidomain_gmsh_reader.hh>
+
 #include <dune/copasi/model/diffusion_reaction.hh>
+#ifndef DUNE_COPASI_SD_LIBRARY
+#include <dune/copasi/model/diffusion_reaction.cc>
+#endif
+
 #include <dune/copasi/model/multidomain_diffusion_reaction.hh>
 #ifndef DUNE_COPASI_MD_LIBRARY
 #include <dune/copasi/model/multidomain_diffusion_reaction.cc>
 #endif
+
 
 #include <dune/grid/multidomaingrid.hh>