+++ CHANGELOG.md
@@ -27,6 +27,11 @@ Types of changes
  -->
 
 ## [Unreleased] ([git-diff][Unreleased-diff])
+### Added
+ - Dune options file receives `CMAKE_OPTIONS` and `MAKE_OPTIONS` !60
+### Fixed
+ - Executables can be compiled without the library !60
+ - All dune cmake flags are now transitively passed to dune-copasi targets !60
 
 ## [1.0.0] ([git-diff][1.0.0-diff]) - 2021-02-11
 ### Added